# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Weather Forecast APIs - SkyAPI Inc.",
    "version": "1.0.0",
    "description": "Reference design documentation of Weather Forecast APIs based on the OpenAPI 3.0 specification. Created by Rohit Parida",
    "termsOfService": "https://skyweather.net/terms",
    "contact": {
      "email": "contact@skyweather.net",
      "name": "Contact SkyAPI Inc.",
      "url": "https://skyweather.net/contacts"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "localhost:8080",
      "variables": {},
      "description": "for testing on local computer"
    }
  ],
...
  "paths": {
    "/v1/locations": {
      "summary": "Operations for managing cities and regions around the world (location)",
      "description": "Clients use these APIs to manage locations available for weather forecast",
      "get": {
        "summary": "Returns a list of managed locations - available for weather forecast",
        "description": "Clients use this API to get a list of managed locations. Each location is uniquely identified by location code, for example 'LACA_US' represents Los Angeles city in California state, in the United States of America (US)",
        "operationId": "getLocations",
        "responses": {
          "200": {
            "description": "successful operation. There are managed locations available.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                },
                "examples": {
                  "A list of locations": {
                    "value": [
                      {
                        "code": "LACA_US",
                        "city_name": "Los Angeles",
                        "region_name": "California",
                        "country_code": "US",
                        "country_name": "United States of America",
                        "enabled": true
                      },
                      {
                        "code": "DELHI_IN",
                        "city_name": "Delhi",
                        "region_name": "Delhi",
                        "country_code": "IN",
                        "country_name": "India",
                        "enabled": false
                      },
                      {
                        "code": "HANOI_VN",
                        "city_name": "Hanoi",
                        "region_name": "Hanoi",
                        "country_code": "VN",
                        "country_name": "Vietnam",
                        "enabled": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "204": {
            "description": "No managed locations available"
          }
        },
        "tags": [
          "Location"
        ]
      },
      "post": {
        "summary": "Adds a new location to be managed for weather forecast",
        "description": "Clients use this API to put a new location into the system. Location code must be manually specified. Country code and country name based on ISO 3166",
        "operationId": "addLocation",
        "requestBody": {
          "description": "a JSON object that represents a location",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "unique code of a location",
                    "example": "LACA_US"
                  },
                  "city_name": {
                    "type": "string",
                    "description": "name of the city",
                    "example": "Los Angeles"
                  },
                  "region_name": {
                    "type": "string",
                    "description": "name of the region, state or province",
                    "example": "California"
                  },
                  "country_code": {
                    "type": "string",
                    "description": "2 characters representing country code based on ISO 3166 standard",
                    "example": "US"
                  },
                  "country_name": {
                    "type": "string",
                    "description": "name of the country based on ISO 3166 standard",
                    "example": "United States of America"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "whether the location enabled (true) or disabled (false)",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful add operation. Location added. Returns details of the newly added location.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "unique code of a location",
                      "example": "LACA_US"
                    },
                    "city_name": {
                      "type": "string",
                      "description": "name of the city",
                      "example": "Los Angeles"
                    },
                    "region_name": {
                      "type": "string",
                      "description": "name of the region, state or province",
                      "example": "California"
                    },
                    "country_code": {
                      "type": "string",
                      "description": "2 characters representing country code based on ISO 3166 standard",
                      "example": "US"
                    },
                    "country_name": {
                      "type": "string",
                      "description": "name of the country based on ISO 3166 standard",
                      "example": "United States of America"
                    },
                    "enabled": {
                      "type": "boolean",
                      "description": "whether the location enabled (true) or disabled (false)",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed add operation. Request body rejected due to some fields have invalid values"
          }
        },
        "tags": [
          "Location"
        ]
      },
      "put": {
        "summary": "Updates an existing location",
        "description": "Clients use this API to modify information of a specific location. Note that location code cannot be changed.",
        "operationId": "updateLocation",
        "requestBody": {
          "description": "a JSON object that represents the location information that needs to be updated",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "unique code of a location",
                    "example": "LACA_US"
                  },
                  "city_name": {
                    "type": "string",
                    "description": "name of the city",
                    "example": "Los Angeles"
                  },
                  "region_name": {
                    "type": "string",
                    "description": "name of the region, state or province",
                    "example": "California"
                  },
                  "country_code": {
                    "type": "string",
                    "description": "2 characters representing country code based on ISO 3166 standard",
                    "example": "US"
                  },
                  "country_name": {
                    "type": "string",
                    "description": "name of the country based on ISO 3166 standard",
                    "example": "United States of America"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "whether the location enabled (true) or disabled (false)",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful update operation. Location updated. Returns details of the recently updated location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "unique code of a location",
                      "example": "LACA_US"
                    },
                    "city_name": {
                      "type": "string",
                      "description": "name of the city",
                      "example": "Los Angeles"
                    },
                    "region_name": {
                      "type": "string",
                      "description": "name of the region, state or province",
                      "example": "California"
                    },
                    "country_code": {
                      "type": "string",
                      "description": "2 characters representing country code based on ISO 3166 standard",
                      "example": "US"
                    },
                    "country_name": {
                      "type": "string",
                      "description": "name of the country based on ISO 3166 standard",
                      "example": "United States of America"
                    },
                    "enabled": {
                      "type": "boolean",
                      "description": "whether the location enabled (true) or disabled (false)",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed update operation. Request body rejected due to some fields have invalid values"
          },
          "404": {
            "description": "No location found with the given code."
          }
        },
        "tags": [
          "Location"
        ]
      }
    },
    "/v1/locations/{code}": {
      "summary": "Operations for managing a specific location identified by location code",
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "schema": {
            "type": "string",
            "example": "NYC_US"
          },
          "required": true,
          "description": "code of the location"
        }
      ],
      "get": {
        "summary": "Returns details of a location",
        "description": "Clients use this API to find an existing location in the database by a specific code",
        "operationId": "getLocation",
        "responses": {
          "200": {
            "description": "Successful retrieval operation. Location found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "unique code of a location",
                      "example": "LACA_US"
                    },
                    "city_name": {
                      "type": "string",
                      "description": "name of the city",
                      "example": "Los Angeles"
                    },
                    "region_name": {
                      "type": "string",
                      "description": "name of the region, state or province",
                      "example": "California"
                    },
                    "country_code": {
                      "type": "string",
                      "description": "2 characters representing country code based on ISO 3166 standard",
                      "example": "US"
                    },
                    "country_name": {
                      "type": "string",
                      "description": "name of the country based on ISO 3166 standard",
                      "example": "United States of America"
                    },
                    "enabled": {
                      "type": "boolean",
                      "description": "whether the location enabled (true) or disabled (false)",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No location found with the given code."
          }
        },
        "tags": [
          "Location"
        ]
      },
      "delete": {
        "summary": "Removes an existing location found by a specific code",
        "description": "Clients use this API to delete a specific location from database",
        "operationId": "deleteLocation",
        "responses": {
          "204": {
            "description": "Successful delete operation. Location removed."
          },
          "404": {
            "description": "No location found with the given code."
          }
        },
        "tags": [
          "Location"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Location": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "unique code of a location",
            "example": "LACA_US"
          },
          "city_name": {
            "type": "string",
            "description": "name of the city",
            "example": "Los Angeles"
          },
          "region_name": {
            "type": "string",
            "description": "name of the region, state or province",
            "example": "California"
          },
          "country_code": {
            "type": "string",
            "description": "2 characters representing country code based on ISO 3166 standard",
            "example": "US"
          },
          "country_name": {
            "type": "string",
            "description": "name of the country based on ISO 3166 standard",
            "example": "United States of America"
          },
          "enabled": {
            "type": "boolean",
            "description": "whether the location enabled (true) or disabled (false)",
            "example": true
          }
        }
      },
...
    },
...
  }
}